backend/recipes/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import Ingredient


class IngredientSerializer(ModelSerializer):
    """Сериализатор для ингредиентов."""

    class Meta:
        model = Ingredient
        fields = ('id', 'name', 'measurement_unit')
